decode: Add scalars args to FieldRootBitBuf
Also move *Value.TryScalarFn to value.go
